emoji_mirror/requirements.txt
# Emoji Mirror Web Interface Requirements

# Core dependencies
streamlit>=1.28.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# PyTorch (for model inference)
torch>=2.0.0
torchvision>=0.15.0

# Note: Other dependencies (models, utils) are in parent project requirements.txt

